refactor(search-facade): the facade is now binded to a searchId
It's responsible to interact with just a slice of the state: the search matching its searchId
The facade must be init with a searchId, all observables are created during the init, and not during the instanciation
because we need to know the searchId.
This way, using the facade will be transparent for other services and component.

This service is not instanciated at Root level and is not meant to be a
singleton, we MUST implicitly inject and instanciate this service for
each searches