feat(elastic); improve link mapping by making sure links are valid
An invalid link will be marked as such. Links *must* have at least a URL
in http, https or ftp. The link name is either a layer name or a file name
inferred from the URL, but is not mandatory.

Also we don't separate the links by nature in the mapper anymore,
this will be done later on