feat(data-fetcher): improve logic for guessing CSV delimiter
Instead of relying on the Papaparse guessing (which fails in an identified
case) we do our own by parsing the first line and using the delimiter there
for the rest of the file.

This notably fixes the issue where a CSV would contain geometries which means
LOTS of commas, and Papaparse would then identify that commas are in fact the
delimiter.