feat(record): disable error report in downloads list (for now)
The logic here is that this error is useless to the user. On the data preview
components, showing an error is important because the user may expect to see
some data and receive only an empty table/map, but for the downloads list
the user has probably no expectation. Even if querying the WFS fails, the user
does not really need to know it. The platform admin might be interested though,
but that is a different use case.