chore: remove heuristics linked to ESRI catalogs with WFS outputs
Previously links with ESRI:REST protocols and "WFSServer" in the URL were
considered WFS services (which they are); we're now relying on the fact
that the harvester will correctly enter the correct OGC:WFS protocol, and
thus we don't need all thos heuristics anymore.

Also removed automatic inferrence of a link "name" property based on a
file name, this prevented links from having empty "name" properties which we
can now rely on to automatically detect the feature type.