fix(data-fetcher): store final request response in cache
Without this once responses were cached they would always fail to be
reused (because a Response object is not serializable)

Storing array buffers (used for excel parsing) is a bit tricky since an
array buffer is not serializable, so I had to convert it to a number array
and then back to a buffer.

Adjusted the useCache mock in the data-fetcher tests to reproduce the original
issue.