feat(repository): add methods to repository for opening  and saving records
Opening a record for edition returns not only the record but also its source,
and whether or not it was already saved once before. If not, it means the
record only exists as a draft.

The record source can be provided when saving a record, thus ensuring
that an existing record XML is not entirely lost when saving.